other half of the checkin
